requirements.txt
discord.py
openai==0.28
yt_dlp
youtube_dl
asyncio
PyNaCl